Rented movies

+++ VideoRentalShop/src/main/java/com.videorental/controllers/movie/MovieEditController.java
@@ -18,6 +18,7 @@ class MovieEditController {
     private static final String M_EDIT_FORM = "movieForm";
     private static final String P_MOVIE_ID = "id";
     private static final String V_MOVIE_EDIT = "movie-edit-view";
+    private static final String V_ERROR_RENTED = "error-rented-view";
     private static final String MOVIE_LIST_URL ="/movie-list";
 
     private final MovieService movieService;
@@ -27,8 +28,8 @@ class MovieEditController {
 
     @Autowired
     private MovieEditController(MovieService movieService,
-       MovieFormFactory movieFormFactory,
-       MovieValidator movieValidator) {
+        MovieFormFactory movieFormFactory,
+        MovieValidator movieValidator) {
 
         Assert.notNull(movieService, "movieService must not be null");
         Assert.notNull(movieFormFactory, "movieFormFactory must not be null");
@@ -38,6 +39,7 @@ class MovieEditController {
         this.movieFormFactory = movieFormFactory;
         this.movieValidator = movieValidator;
     }
+
 
     @InitBinder(M_EDIT_FORM)
     private void initBinder(WebDataBinder binder) {
@@ -56,6 +58,12 @@ class MovieEditController {
         return movieFormFactory.create(id);
     }
 
+    @ModelAttribute(M_EDIT_FORM)
+    private boolean isMovieRented(Long id) {
+
+        return movieService.isMovieRented(id);
+    }
+
     @GetMapping
     private String showMovieForm() {
 
@@ -64,7 +72,7 @@ class MovieEditController {
 
     @PostMapping
     private String addOrUpdateMovie(@ModelAttribute(M_EDIT_FORM) @Valid MovieForm movieForm,
-                                       BindingResult bindingResult) {
+        BindingResult bindingResult) {
 
         if (bindingResult.hasErrors()) {
 
@@ -79,8 +87,13 @@ class MovieEditController {
     @GetMapping("/delete")
     private String deleteEmployee(@RequestParam(P_MOVIE_ID) Long id) {
 
-        movieService.deleteById(id);
+        boolean movieRented = isMovieRented(id);
 
+        if (movieRented) {
+            return V_ERROR_RENTED;
+        }
+
+        movieService.deleteById(id);
         return "redirect:" + MOVIE_LIST_URL;
     }
 }
